+++ stdlib/VERSIONS
@@ -19,6 +19,7 @@
 
 __future__: 3.0-
 __main__: 3.0-
+_asyncio: 3.0-
 _ast: 3.0-
 _bisect: 3.0-
 _bootlocale: 3.4-3.9
